garde_backend/requirements.txt
# cffi==1.16.0
# chardet==5.2.0
# cssselect2==0.7.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
# pillow==10.3.0
# pycparser==2.22
# pyphen==0.15.0
python-dateutil==2.9.0.post0
# six==1.16.0
sqlparse==0.5.0
# tinycss2==1.3.0
# webencodings==0.5.1
# zopfli==0.2.3
pytz==2024.1
